🔨 refactor legends dropping manager pattern
