+++ packages/@ourworldindata/grapher/src/barCharts/DiscreteBarChart.tsx
@@ -33,10 +33,9 @@ import {
     SizedDiscreteBarSeries,
 } from "./DiscreteBarChartConstants"
 import { CategoricalBin, ColorScaleBin } from "../color/ColorScaleBin"
-import {
-    HorizontalColorLegendManager,
-    HorizontalNumericColorLegend,
-} from "../legend/HorizontalColorLegends"
+import { HorizontalNumericColorLegend } from "../legend/HorizontalNumericColorLegend"
+import { HorizontalNumericColorLegendState } from "../legend/HorizontalNumericColorLegendState"
+import { ExternalColorLegendData } from "../legend/HorizontalColorLegendTypes"
 import { DiscreteBarChartState } from "./DiscreteBarChartState"
 import { ChartComponentProps } from "../chart/ChartTypeMap.js"
 import {
@@ -73,7 +72,7 @@ export type DiscreteBarChartProps = ChartComponentProps<DiscreteBarChartState>
 @observer
 export class DiscreteBarChart
     extends React.Component<DiscreteBarChartProps>
-    implements ChartInterface, AxisManager, HorizontalColorLegendManager
+    implements ChartInterface, AxisManager
 {
     base = React.createRef<SVGGElement>()
 
@@ -526,9 +525,15 @@ export class DiscreteBarChart
     }
 
     private renderLegend(): React.ReactElement | null {
-        if (!this.showColorLegend) return null
+        if (!this.showColorLegend || !this.numericLegend) return null
 
-        return <HorizontalNumericColorLegend manager={this} />
+        return (
+            <HorizontalNumericColorLegend
+                state={this.numericLegend}
+                x={this.legendX}
+                y={this.numericLegendY}
+            />
+        )
     }
 
     private renderAxis(): React.ReactElement {
@@ -599,20 +604,12 @@ export class DiscreteBarChart
         return this.hasColorLegend && !!this.manager.showLegend
     }
 
-    @computed get legendX(): number {
+    @computed private get legendX(): number {
         return this.bounds.x
     }
 
-    @computed get legendMaxWidth(): number {
-        return this.bounds.width
-    }
-
-    @computed get legendAlign(): HorizontalAlign {
-        return HorizontalAlign.center
-    }
-
     // TODO just pass colorScale to legend and let it figure it out?
-    @computed get numericLegendData(): ColorScaleBin[] {
+    @computed private get numericLegendData(): ColorScaleBin[] {
         const legendBins = this.chartState.colorScale.legendBins.slice()
 
         // Show a "Projected data" legend item with a striped pattern if appropriate
@@ -642,9 +639,9 @@ export class DiscreteBarChart
     }
 
     // Used when the bars are colored by a numeric scale
-    numericLegendStyleConfig = NUMERIC_LEGEND_STYLE
+    private readonly numericLegendStyleConfig = NUMERIC_LEGEND_STYLE
 
-    @computed get externalLegend(): HorizontalColorLegendManager | undefined {
+    @computed get externalLegend(): ExternalColorLegendData | undefined {
         if (this.hasColorLegend) {
             return {
                 numericLegendData: this.numericLegendData,
@@ -654,25 +651,33 @@ export class DiscreteBarChart
         return undefined
     }
 
-    @computed get numericBinSize(): number {
+    @computed private get numericBinSize(): number {
         return 0.625 * this.fontSize
     }
 
-    legendTickSize = 1
+    private readonly legendTickSize = 1
 
     @computed private get numericLegend():
-        | HorizontalNumericColorLegend
+        | HorizontalNumericColorLegendState
         | undefined {
         return this.manager.showLegend
-            ? new HorizontalNumericColorLegend({ manager: this })
+            ? new HorizontalNumericColorLegendState(this.numericLegendData, {
+                  fontSize: this.fontSize,
+                  legendMaxWidth: this.bounds.width,
+                  legendAlign: HorizontalAlign.center,
+                  legendTitle: this.legendTitle,
+                  legendTickSize: this.legendTickSize,
+                  numericBinSize: this.numericBinSize,
+                  numericLegendStyleConfig: this.numericLegendStyleConfig,
+              })
             : undefined
     }
 
-    @computed get numericLegendY(): number {
+    @computed private get numericLegendY(): number {
         return this.bounds.top
     }
 
-    @computed get legendTitle(): string | undefined {
+    @computed private get legendTitle(): string | undefined {
         return this.chartState.hasColorScale
             ? this.chartState.colorScale.legendDescription
             : undefined
