+++ packages/@ourworldindata/grapher/src/mapCharts/MapChart.tsx
@@ -8,11 +8,11 @@ import {
 } from "@ourworldindata/utils"
 import { observable, computed, action, makeObservable } from "mobx"
 import { observer } from "mobx-react"
-import {
-    HorizontalCategoricalColorLegend,
-    HorizontalColorLegendManager,
-    HorizontalNumericColorLegend,
-} from "../legend/HorizontalColorLegends"
+import { HorizontalNumericColorLegend } from "../legend/HorizontalNumericColorLegend"
+import { HorizontalCategoricalColorLegend } from "../legend/HorizontalCategoricalColorLegend"
+import { HorizontalNumericColorLegendState } from "../legend/HorizontalNumericColorLegendState"
+import { HorizontalCategoricalColorLegendState } from "../legend/HorizontalCategoricalColorLegendState"
+import { ExternalColorLegendData } from "../legend/HorizontalColorLegendTypes"
 
 import { MapTooltip } from "./MapTooltip"
 import { TooltipState } from "../tooltip/Tooltip.js"
@@ -79,10 +79,7 @@ export const MAP_LEGEND_MAX_WIDTH_RATIO = 0.95
 @observer
 export class MapChart
     extends Component<MapChartProps>
-    implements
-        ChartInterface,
-        HorizontalColorLegendManager,
-        ChoroplethMapManager
+    implements ChartInterface, ChoroplethMapManager
 {
     constructor(props: MapChartProps) {
         super(props)
@@ -227,21 +224,13 @@ export class MapChart
         }
     }
 
-    @computed get externalLegend(): HorizontalColorLegendManager | undefined {
-        const {
-            numericLegendData,
-            categoricalLegendData,
-            legendMaxWidth,
-            legendStyleConfig,
-        } = this
-
+    @computed get externalLegend(): ExternalColorLegendData | undefined {
         if (this.manager.showLegend) return undefined
 
         return {
-            numericLegendData,
-            categoricalLegendData,
-            legendMaxWidth,
-            legendStyleConfig,
+            numericLegendData: this.numericLegendData,
+            categoricalLegendData: this.categoricalLegendData,
+            legendStyleConfig: this.legendStyleConfig,
         }
     }
 
@@ -496,51 +485,69 @@ export class MapChart
     }
 
     @computed private get numericLegendHeight(): number {
-        return this.numericLegend ? this.numericLegend.height : 0
+        return this.numericLegendState?.height ?? 0
     }
 
     @computed private get categoryLegendHeight(): number {
-        return this.categoryLegend ? this.categoryLegend.height : 0
+        return this.categoryLegendState?.height ?? 0
     }
 
-    @computed private get categoryLegend():
-        | HorizontalCategoricalColorLegend
+    @computed private get categoryLegendState():
+        | HorizontalCategoricalColorLegendState
         | undefined {
-        return this.manager.showLegend && this.categoricalLegendData.length > 1
-            ? new HorizontalCategoricalColorLegend({ manager: this })
-            : undefined
+        if (!this.manager.showLegend || this.categoricalLegendData.length <= 1)
+            return undefined
+        return new HorizontalCategoricalColorLegendState(
+            this.categoricalLegendData,
+            {
+                fontSize: this.fontSize,
+                legendWidth: this.legendMaxWidth,
+                legendAlign: this.legendAlign,
+                resolveLegendBinEmphasis: (bin) =>
+                    this.resolveLegendBinEmphasis(bin),
+                legendStyleConfig: this.legendStyleConfig,
+            }
+        )
     }
 
-    @computed private get numericLegend():
-        | HorizontalNumericColorLegend
+    @computed private get numericLegendState():
+        | HorizontalNumericColorLegendState
         | undefined {
-        return this.manager.showLegend && this.numericLegendData.length > 1
-            ? new HorizontalNumericColorLegend({ manager: this })
-            : undefined
+        if (!this.manager.showLegend || this.numericLegendData.length <= 1)
+            return undefined
+        return new HorizontalNumericColorLegendState(this.numericLegendData, {
+            fontSize: this.fontSize,
+            legendMaxWidth: this.legendMaxWidth,
+            legendAlign: this.legendAlign,
+            numericBinSize: this.numericBinSize,
+            resolveLegendBinEmphasis: (bin) =>
+                this.resolveLegendBinEmphasis(bin),
+            legendStyleConfig: this.legendStyleConfig,
+        })
     }
 
-    @computed get categoryLegendY(): number {
-        if (!this.categoryLegend) return 0
+    @computed private get categoryLegendY(): number {
+        if (!this.categoryLegendState) return 0
 
         return (
             this.bounds.bottom -
-            this.categoryLegend.height -
+            this.categoryLegendState.height -
             PADDING_BELOW_MAP_LEGEND
         )
     }
 
-    @computed get legendAlign(): HorizontalAlign {
+    @computed private get legendAlign(): HorizontalAlign {
         return HorizontalAlign.center
     }
 
-    @computed get numericLegendY(): number {
-        if (!this.numericLegend) return 0
+    @computed private get numericLegendY(): number {
+        if (!this.numericLegendState) return 0
         return (
             this.bounds.bottom -
             this.numericLegendHeight -
             PADDING_BELOW_MAP_LEGEND -
             // If present, the category legend is placed below the numeric legend
-            (this.categoryLegend
+            (this.categoryLegendState
                 ? this.categoryLegendHeight + PADDING_BETWEEN_MAP_LEGENDS
                 : 0)
         )
@@ -571,15 +578,30 @@ export class MapChart
     }
 
     renderMapLegend(): React.ReactElement {
-        const { numericLegend, categoryLegend } = this
+        const { numericLegendState, categoryLegendState } = this
 
         return (
             <>
-                {numericLegend && (
-                    <HorizontalNumericColorLegend manager={this} />
+                {numericLegendState && (
+                    <HorizontalNumericColorLegend
+                        state={numericLegendState}
+                        x={this.legendX}
+                        y={this.numericLegendY}
+                        onMouseEnter={this.onLegendMouseEnter}
+                        onMouseLeave={this.onLegendMouseLeave}
+                        onMouseOver={this.onLegendMouseOver}
+                    />
                 )}
-                {categoryLegend && (
-                    <HorizontalCategoricalColorLegend manager={this} />
+                {categoryLegendState && (
+                    <HorizontalCategoricalColorLegend
+                        state={categoryLegendState}
+                        x={this.legendX}
+                        y={this.categoryLegendY}
+                        onMouseEnter={this.onLegendMouseEnter}
+                        onMouseLeave={this.onLegendMouseLeave}
+                        onMouseOver={this.onLegendMouseOver}
+                        isStatic={this.isStatic}
+                    />
                 )}
             </>
         )
