+++ packages/@ourworldindata/grapher/src/slopeCharts/SlopeChart.tsx
@@ -63,7 +63,7 @@ import {
 import { TooltipFooterIcon } from "../tooltip/TooltipProps"
 
 import { Halo } from "@ourworldindata/components"
-import { HorizontalColorLegendManager } from "../legend/HorizontalColorLegends"
+import { ExternalColorLegendData } from "../legend/HorizontalColorLegendTypes"
 import { CategoricalBin } from "../color/ColorScaleBin"
 import {
     GRAPHER_BACKGROUND_DEFAULT,
@@ -321,7 +321,7 @@ export class SlopeChart
         }).width
     }
 
-    @computed get externalLegend(): HorizontalColorLegendManager | undefined {
+    @computed get externalLegend(): ExternalColorLegendData | undefined {
         if (!this.manager.showSeriesLabels) {
             const categoricalLegendData = this.series.map(
                 (series, index) =>
