+++ packages/@ourworldindata/grapher/src/stackedCharts/MarimekkoChart.tsx
@@ -48,10 +48,8 @@ import {
     makeTooltipRoundingNotice,
     makeTooltipToleranceNotice,
 } from "../tooltip/Tooltip"
-import {
-    HorizontalCategoricalColorLegend,
-    HorizontalColorLegendManager,
-} from "../legend/HorizontalColorLegends"
+import { HorizontalCategoricalColorLegend } from "../legend/HorizontalCategoricalColorLegend"
+import { HorizontalCategoricalColorLegendState } from "../legend/HorizontalCategoricalColorLegendState"
 import { CategoricalBin, ColorScaleBin } from "../color/ColorScaleBin"
 import { LegendStyleConfig } from "../legend/LegendStyleConfig"
 import { Emphasis } from "../interaction/Emphasis"
@@ -82,7 +80,7 @@ export type MarimekkoChartProps = ChartComponentProps<MarimekkoChartState>
 @observer
 export class MarimekkoChart
     extends React.Component<MarimekkoChartProps>
-    implements ChartInterface, HorizontalColorLegendManager, AxisManager
+    implements ChartInterface, AxisManager
 {
     base = React.createRef<SVGGElement>()
 
@@ -289,19 +287,19 @@ export class MarimekkoChart
         return this.legend.height > 0 ? this.fontSize : 0
     }
 
-    @computed get legendX(): number {
+    @computed private get legendX(): number {
         return this.bounds.x
     }
 
-    @computed get categoryLegendY(): number {
+    @computed private get categoryLegendY(): number {
         return this.bounds.top
     }
 
-    @computed get legendWidth(): number {
+    @computed private get legendWidth(): number {
         return this.bounds.width
     }
 
-    @computed get legendAlign(): HorizontalAlign {
+    @computed private get legendAlign(): HorizontalAlign {
         return HorizontalAlign.left
     }
 
@@ -309,7 +307,7 @@ export class MarimekkoChart
         return this.manager.detailsOrderedByReference ?? []
     }
 
-    @computed get categoricalLegendData(): CategoricalBin[] {
+    @computed private get categoricalLegendData(): CategoricalBin[] {
         const { colorColumnSlug, colorScale, series } = this
         if (colorColumnSlug) {
             return colorScale.categoricalLegendBins
@@ -329,7 +327,7 @@ export class MarimekkoChart
         return []
     }
 
-    resolveLegendBinEmphasis(bin: ColorScaleBin): Emphasis {
+    private resolveLegendBinEmphasis(bin: ColorScaleBin): Emphasis {
         const { focusColorBin } = this
 
         // If nothing is focused, all items are active
@@ -344,7 +342,8 @@ export class MarimekkoChart
         return isFocused ? Emphasis.Highlighted : Emphasis.Muted
     }
 
-    legendStyleConfig: LegendStyleConfig = LEGEND_STYLE_FOR_STACKED_CHARTS
+    private readonly legendStyleConfig: LegendStyleConfig =
+        LEGEND_STYLE_FOR_STACKED_CHARTS
 
     @computed get hoverColors(): string[] {
         if (this.focusColorBin) return [this.focusColorBin.color]
@@ -379,8 +378,18 @@ export class MarimekkoChart
         this.focusColorBin = undefined
     }
 
-    @computed private get legend(): HorizontalCategoricalColorLegend {
-        return new HorizontalCategoricalColorLegend({ manager: this })
+    @computed private get legend(): HorizontalCategoricalColorLegendState {
+        return new HorizontalCategoricalColorLegendState(
+            this.categoricalLegendData,
+            {
+                fontSize: this.fontSize,
+                legendWidth: this.legendWidth,
+                legendAlign: this.legendAlign,
+                resolveLegendBinEmphasis: (bin: ColorScaleBin) =>
+                    this.resolveLegendBinEmphasis(bin),
+                legendStyleConfig: this.legendStyleConfig,
+            }
+        )
     }
 
     @action.bound private onEntityMouseOver(entityName: string): void {
@@ -543,7 +552,14 @@ export class MarimekkoChart
                     backgroundColor={manager.backgroundColor}
                 />
                 {this.showLegend && (
-                    <HorizontalCategoricalColorLegend manager={this} />
+                    <HorizontalCategoricalColorLegend
+                        state={this.legend}
+                        x={this.legendX}
+                        y={this.categoryLegendY}
+                        onMouseOver={this.onLegendMouseOver}
+                        onMouseLeave={this.onLegendMouseLeave}
+                        isStatic={this.isStatic}
+                    />
                 )}
                 {this.renderBars()}
                 {this.labelLines}
